Change parsers to fit new Forecast Excel file format. (#32)
* Updated parsers to the new excel file format.

* Updated tests for SpreadsheetUtils.

* Fixed bug while parsing Packing productivity.

* Fixed bug while parsing Packing productivity.

* Removed unused methods.

+++ application/src/main/java/com/mercadolibre/planning/model/me/gateways/planningmodel/dtos/AreaDistribution.java
@@ -7,5 +7,5 @@ import lombok.Getter;
 @Builder
 public class AreaDistribution {
     private String areaId;
-    private long quantity;
+    private double quantity;
 }
